+++ src/main/java/com/qcloud/Module/Morphling.java
@@ -1,0 +1,11 @@
+package com.qcloud.Module;
+
+public class Morphling extends Base {
+	public Morphling(String module) {
+		serverHost = module + ".api.qcloud.com";
+	}
+
+    public void morph(String module) {
+        serverHost = module + ".api.qcloud.com";
+    }
+}